+++ banking/src/main/java/casino/banking/services/user/UserBalanceService.java
@@ -3,7 +3,8 @@ package casino.banking.services.user;
 import casino.banking.view.user.UserDTO;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface UserBalanceService {
-    UserDTO depositBalanceById(Long userId, BigDecimal amount, int decimals);
+    UserDTO depositBalanceById(Long userId, BigInteger amount, int decimals);
 }
